test(e2e): adapt to password-change revocation and read-only 2FA setup
- settings.spec.ts: changing the password now revokes existing API/device
  tokens (matching seahub), so refresh the admin API token persisted for
  later specs. The acting browser session is deliberately preserved.
- two-factor.spec.ts: setup is a POST now (GET is read-only), and the fresh
  context needs an explicit baseURL.
- clippy: drop a needless borrow and a same-type cast.

+++ server/src/ui/settings.rs
@@ -190,7 +190,7 @@ pub async fn change_password(
     // unlike registration and password reset (M-5).
     if let Err(msg) = crate::service::auth::password::validate_password(
         &form.new_password,
-        state.config.auth.password_min_length as u32,
+        state.config.auth.password_min_length,
         state.config.auth.require_strong_password,
     ) {
         return render_settings_error(&state, &user, Some(msg)).await;
